requirements.txt
streamlit
google-generativeai
googletrans==4.0.0-rc1 